Split blocking activity logger
